# contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/signup": {
      "post": {
        "operationId": "signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupStartedAnswer"
                }
              }
            }
          },
          "400": {
            "description": "Validation, PHONE_INVALID, PHONE_NOT_SUPPORTED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The email or phone is already taken by another account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "PHONE_CODE_RESEND_TOO_SOON (texted), EMAIL_CODE_RESEND_TOO_SOON (mailed), RATE_LIMITED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "SMS_NOT_CONFIGURED, SMS_SEND_FAILED (texted); EMAIL_NOT_CONFIGURED, EMAIL_SEND_FAILED (mailed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "summary": "A3: create the account and send a 6-digit code: texted to the phone by default, mailed to the email when SIGNUP_VERIFY_CHANNEL is email (R-39; the answer then carries `channel: email`). No session yet (R-36). A number another account only typed is not taken from it by a sign-up that proves nothing: a mailed sign-up goes ahead without the number (`phoneNotSaved: true`), and a texted one takes it when its code proves it.",
        "tags": [
          "Auth",
          "app"
        ]
      }
    },
    "/auth/phone/verify/start": {
      "post": {
        "operationId": "phoneVerifyStart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhoneVerifyStartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneCodeSentAnswer"
                }
              }
            }
          },
          "400": {
            "description": "Validation, PHONE_INVALID, PHONE_NOT_SUPPORTED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "SIGNUP_CHANNEL_DISABLED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "PHONE_CODE_RESEND_TOO_SOON, RATE_LIMITED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "SMS_NOT_CONFIGURED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "summary": "A4 (texted sign-up): send another code. The same answer for any number and secret; only a live sign-up is texted. Answers 409 SIGNUP_CHANNEL_DISABLED while codes are mailed.",
        "tags": [
          "Auth",
          "app"
        ]
      }
    },
    "/auth/phone/verify/confirm": {
      "post": {
        "operationId": "phoneVerifyConfirm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhoneVerifyConfirmDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionAnswer"
                }
              }
            }
          },
          "400": {
            "description": "Validation, PHONE_CODE_INVALID (also for a secret that is not live: restart sign-up after the fourth 400 in a row).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "PHONE_ALREADY_CONFIRMED, SIGNUP_CHANNEL_DISABLED, or the email or phone was taken meanwhile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "PHONE_CODE_LOCKED, RATE_LIMITED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "summary": "A4 (texted sign-up): check the texted code (and the mailed one when sign-up asked for it). Right: the first session. Answers 409 SIGNUP_CHANNEL_DISABLED while codes are mailed.",
        "tags": [
          "Auth",
          "app"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionAnswer"
                }
              }
            }
          },
          "400": {
            "description": "Validation: a field missing or of the wrong type.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Wrong password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "EMAIL_NOT_VERIFIED, PHONE_NOT_CONFIRMED (a sign-up whose phone code was never entered: resume it).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "USER_NOT_IN_WAWUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "summary": "A22: sign in by email or phone, and password.",
        "tags": [
          "Auth",
          "app"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairAnswer"
                }
              }
            }
          },
          "400": {
            "description": "Validation: a field missing or of the wrong type.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "summary": "Trade a refresh token for a new pair; the old one stops working.",
        "tags": [
          "Auth",
          "app"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedOutAnswer"
                }
              }
            }
          },
          "400": {
            "description": "Validation: a field missing or of the wrong type.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "summary": "Z2: sign out. The refresh token sent stops working. The same answer for a live, a dead or an unknown token.",
        "tags": [
          "Auth",
          "app"
        ]
      }
    },
    "/auth/change-password": {
      "post": {
        "operationId": "changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairAnswer"
                }
              }
            }
          },
          "400": {
            "description": "Validation (new password under 8 or over 128 characters), CURRENT_PASSWORD_WRONG, PASSWORD_UNCHANGED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "SESSION_INVALID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "PASSWORD_NOT_SET (an account that never had a password).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED (five wrong current passwords in 15 minutes).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "wawu-id": []
          }
        ],
        "summary": "Z2: change the password of the signed-in account. Every other device is signed out; this one gets a new pair to keep using.",
        "tags": [
          "Auth",
          "app"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetRequestedAnswer"
                }
              }
            }
          },
          "400": {
            "description": "Validation: a field missing or of the wrong type, or a `method` other than `sms` or `email`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "summary": "A23: ask for a reset. The same answer whether or not an account exists; a link is mailed only to a proven email.",
        "tags": [
          "Auth",
          "app"
        ]
      }
    },
...
  },
  "info": {
    "title": "WAWU ID",
    "description": "The sign-in service: the routes the mobile app calls.\n\nEvery success body is `{ data: ... }`, exactly as described here (no\ninterceptor). Every error body is ErrorBody. Generated from the code by\n`npm run contract:build`; never edited by hand.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "wawu-id": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "minLength": 7,
            "maxLength": 25,
            "description": "As typed: 0803..., +234... or 234..."
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "occupation": {
            "type": "string",
            "maxLength": 120
          },
          "accountType": {
            "type": "string",
            "enum": [
              "user",
              "creator"
            ],
            "description": "The A2 pick: discover is `user`, earn is `creator`. Without it the sequence treats the account as `user`."
          }
        },
        "required": [
          "email",
          "phone",
          "password"
        ]
      },
      "SignupStartedSchema": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "description": "The number as stored (+234...)."
          },
          "expiresIn": {
            "type": "number",
            "description": "Seconds the code can be used for."
          },
          "resendIn": {
            "type": "number",
            "description": "Seconds until another code can be asked for."
          },
          "attempt": {
            "type": "string",
            "description": "Shown once. Keep it with the sign-up (it survives the app closing) and send it on every code and resume call."
          },
          "emailCodeRequired": {
            "type": "boolean",
            "description": "Texted sign-up only: the confirm call must also carry `emailCode`, the code mailed to the email. Always false when the code was mailed."
          },
          "channel": {
            "type": "string",
            "enum": [
              "email"
            ],
            "description": "`email`: the code was mailed (`SIGNUP_VERIFY_CHANNEL=email`, DECISIONS R-39): send it to `signup/email-code/*`. Absent: the code was texted (the default): `phone/verify/*`."
          },
          "maskedEmail": {
            "type": "string",
            "description": "With `channel: email`: where the code went, written a•••@example.com. Show it on A4."
          },
          "phoneNotSaved": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "With `channel: email`, present (true) only when the phone number typed was NOT saved to the account because another account holds it. The account is made without a phone; say so on A4 in plain words. The person adds and proves a phone later (BACKEND_GAPS G-222)."
          }
        },
        "required": [
          "phone",
          "expiresIn",
          "resendIn",
          "attempt",
          "emailCodeRequired"
        ]
      },
      "SignupStartedAnswer": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SignupStartedSchema"
          }
        },
        "required": [
          "data"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "code": {
            "type": "string",
            "enum": [
              "USER_NOT_IN_WAWUID",
              "EMAIL_NOT_VERIFIED",
              "PHONE_NOT_CONFIRMED",
              "PHONE_INVALID",
              "PHONE_NOT_SUPPORTED",
              "PHONE_CODE_INVALID",
              "PHONE_CODE_LOCKED",
              "PHONE_CODE_RESEND_TOO_SOON",
              "PHONE_ALREADY_CONFIRMED",
              "RATE_LIMITED",
              "SMS_NOT_CONFIGURED",
              "SMS_SEND_FAILED",
              "SESSION_INVALID",
              "SIGNUP_STEP_OUT_OF_ORDER",
              "SIGNUP_ALREADY_FINISHED",
              "EMAIL_NOT_SET",
              "EMAIL_ALREADY_PROVEN",
              "EMAIL_CODE_INVALID",
              "EMAIL_CODE_LOCKED",
              "EMAIL_CODE_RESEND_TOO_SOON",
              "EMAIL_ALREADY_CONFIRMED",
              "EMAIL_NOT_CONFIGURED",
              "EMAIL_SEND_FAILED",
              "SIGNUP_CHANNEL_DISABLED",
              "CURRENT_PASSWORD_WRONG",
              "PASSWORD_UNCHANGED",
              "PASSWORD_NOT_SET"
            ],
            "description": "Machine-readable cause, when there is one. A body without it is a plain refusal (a 400 from validation, a 401 wrong password, a 409 taken email or phone)."
          },
          "message": {
            "type": "string",
            "description": "A sentence for people. Validation errors join theirs with \", \"."
          },
          "retryAfterSeconds": {
            "type": "number",
            "description": "Seconds to wait, with every 429 (also sent as a Retry-After header)."
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
      "PhoneVerifyStartDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "minLength": 7,
            "maxLength": 25
          },
          "attempt": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "The secret `POST /auth/signup` returned for this sign-up."
          }
        },
        "required": [
          "phone",
          "attempt"
        ]
      },
...
      "PhoneCodeSentAnswer": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PhoneCodeSentSchema"
          }
        },
        "required": [
          "data"
        ]
      },
      "PhoneVerifyConfirmDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "minLength": 7,
            "maxLength": 25
          },
          "attempt": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "The secret `POST /auth/signup` returned for this sign-up."
          },
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$"
          },
          "emailCode": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "description": "Required when sign-up answered `emailCodeRequired: true`."
          }
        },
        "required": [
          "phone",
          "attempt",
          "code"
        ]
      },
...
      "SessionAnswer": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Session"
          }
        },
        "required": [
          "data"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "description": "Email, or phone as typed (0803... or +234...)."
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "identifier",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "TokenPairAnswer": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TokenPairSchema"
          }
        },
        "required": [
          "data"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "The refresh token this device holds."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "SignedOutAnswer": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SignedOutSchema"
          }
        },
        "required": [
          "data"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "The password the account has now."
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "description": "The same length rule as a reset (8 to 128 characters)."
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "description": "Email or phone number."
          },
          "method": {
            "type": "string",
            "enum": [
              "sms",
              "email"
            ],
            "description": "Without it, `sms`. The app sends `email`."
          }
        },
        "required": [
          "identifier"
        ]
      },
...
      "ResetRequestedAnswer": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ResetRequested"
          }
        },
        "required": [
          "data"
        ]
      },
...
    }
  }
}